Hotfix integer under flow

+++ gsmgogo-api/src/main/java/team/gsmgogo/domain/bet/controller/dto/BetRequest.java
@@ -18,7 +18,7 @@ public class BetRequest {
     private Long matchId;
     
     @NotNull
-    private Long betPoint;
+    private Integer betPoint;
 
     @NotNull
     @JsonProperty("team_a_score")
